connect to gevent server through redis with ssl and call commands

+++ JumpScale9RecordChain/servers/gedis/GedisServer.py
@@ -12,7 +12,7 @@ import imp
 TEMPLATE = """
 addr = "localhost"
 port = "9900"
-# ssl = false
+ssl = false
 adminsecret_ = ""
 """
 JSConfigBase = j.tools.configmanager.base_class_config
@@ -36,10 +36,6 @@ class GedisServer(StreamServer, JSConfigBase):
             self.logger.info("ssl enabled, keys in %s"%self.ssl_priv_key_path)
             self.sslkeys_generate()
 
-            #############7ossam
-            # there was no session here on tmux, happened first, come here again call it three
-            #4 this appeared on tmux, again on tmux call it 5
-            # outside of tmux after failure this appeared
             self.server = StreamServer(
                 (host, port), spawn=Pool(), handle=self.__handle_connection, keyfile=self.ssl_priv_key_path, certfile=self.ssl_cert_path)
         else:
@@ -75,29 +71,38 @@ class GedisServer(StreamServer, JSConfigBase):
         # import ipdb;ipdb.set_trace()
         self.logger.info("add cmd %s" % cmd)
         content = inspect.getsource(callback)
-        j.sal.fs.writeFile(self._cmds_path, contents='\n'+content, append=True)
+        lines = content.splitlines()
+        content = ""
+        for line in lines:
+            line = line.replace("self,", "")
+            content = content + line[4:] + "\n"
+        content
 
+        if not j.sal.fs.exists(path=self._cmds_path):
+            j.sal.fs.writeFile(self._cmds_path, contents=content, append=True)
+        else:
+            __cmds = imp.load_source(name="cmds.py", pathname=self._cmds_path)
+            if cmd+"_cmd" not in __cmds.__dir__():
+                j.sal.fs.writeFile(self._cmds_path, contents='\n' + content, append=True)
 
     def __handle_connection(self, socket, address):
-        # import ipdb; ipdb.set_trace()
         self.logger.info('connection from {}'.format(address))
         parser = CommandParser(socket)
         response = ResponseWriter(socket)
-
-        self._cmds = imp.load_source(pathname=self._cmds_path)
+        self._cmds = imp.load_source(name="cmds.py", pathname=self._cmds_path)
 
         try:
             while True:
                 request = parser.read_request()
                 cmd = request[0]
-                if cmd not in self._cmds.__dir__():
+                if cmd.decode("utf-8")+"_cmd".lower() not in (command.lower() for command in self._cmds.__dir__()):
                     response.error('command not supported')
                     continue
 
                 # execute command callback
                 result = ""
                 try:
-                    result = getattr(self._cmds, cmd)(request)
+                    result = getattr(self._cmds, cmd.decode("utf-8")+"_cmd")(request)
                     self.logger.debug(
                         "Callback done and result {} , type {}".format(result, type(result)))
                 except Exception as e:
